Remove duplicate ReactMapTooltip file
